mirror/requirements.txt
opentelemetry-proto